MOV-37 Created a scrollable Screening List for the Admin landing page.
Currently contains bogus data, but will be rewritten to work with real data once the Screening table in the database will be populated and linked to it.

The UI for the list is quite simple, but it has all the necessary buttons, which are not yet implemented. Uses the custom List cell.