MOV-6 Added code for inserting new screenings into the database.
The Add Screening form is operational now. It still needs tweaking, as it allows for screenings in the past to be created, but it mainly works in both cases - adding screening for new movie and for already existing movie.

Basic input correctness checks are treated by the code - empty fields, duplicate movies (this one is treated automatically without notifying the user). Also, the fields for a new movie aren't considered when on the "Add Existing..." tab and vice-versa.

Unit tests have been disabled once again.