MOV-2 Adapted the changePassword() from the change pass be4 login story.
The method signature was changed to be used in both cases - password change before login and password change after login.

It was also overloaded by a method which changes the password after login - the new method still uses the overloaded one.

Tweaked the ChangePasswordBeforeLoginController to work with the changes from UserService.changeService().