MOV-12 Cancelled screenings now display accordingly.
Because the screenings list is populated from the database when the Scene is loaded, deleted screenings will not disappear until the page is refreshed. To prevent unwanted scenarios caused by this chink, the ScreeningListCell was modified to disable the buttons correspoding to the deleted screening instantly.

After refreshing, the deleted screening disappears just as intended.