MOV-12 Reworked the mailing methods in CommService.
In this story, users should be notified via e-mail when one of the movie screenings they have bookings for has been cancelled. In this sense, I wrote a method for sending an email to multiple recipients. The common code between this and the old email function was factored into createMailDraft().

Also changed the e-mail address the app uses to send notifications because the old one was just for fun.