MOV-11 Moved the ScreeningListCell class.
For this story, the "Bookings" buttons in the Screening Lists need to be linked to the Bookings page. This wasn't possible with ScreeningListCell as a separate class, not extending AbstractController (because of the private access to changeScene()). 

For solving this problem, ScreeningListCell was added to the MainMenuAdminController class as an Inner Class. This better suits the single purpose of ScreeningListCell, making it inaccessible from the outside.