MOV-108 Added tests for all the current methods in ScreeningService.
Found a bug regarding addScreening(), which occured when adding a screening for an existing movie using the form for new movies. Instead of using the movie length data from the database to check if the time frame for the screening is free, the system used the possibly incorrect length provided via the form. This was solved by swapping the positions of two code blocks in addScreening().

Also deleted the unused methods findAllScreeningsForCinema().